Adds feature to check if random_string exists
This feature is in the random_string function.It searches the
database for the generated string. If that operation returns a value
not equal to 'none', the random string function creates a new string
and searches for it in the database until there is no match and then
returns that value. This process is looped. The purpose of this feature
is to ensure the generated string for two different urls is never the
same.