Adds feature to check if long_url exists
This feature is within the add_url function for the '/url' route.
The feature checks if the given long_url from the user already exists
in the database. If the url exists, we query the db with the long_url
and return the existing short_url, rather than create a new one.